Sorting warning in numvec

+++ src/numvec.hpp
@@ -105,7 +105,7 @@ std::vector<T> to_vec(py::array_t<T,1>& tin)
 {
 	numvec<double> in(tin);
 	std::vector<T> out(in.size());
-	for(int i=0;i<in.size(); ++i)
+	for(size_t i=0;i<in.size(); ++i)
 		out[i] = in[i];
 	return out;
 }
